hm-done-reborne-4

+++ yatube/posts/views.py
@@ -132,10 +132,10 @@ def profile_follow(request, username):
 
 @login_required
 def profile_unfollow(request, username):
-    Follow.objects.filter(
-        user=request.user,
-        author=get_object_or_404(User, username=username)).exists()
-    Follow.objects.filter(
-        user=request.user,
-        author=get_object_or_404(User, username=username)).delete()
-    return redirect('posts:profile', username)
+    if Follow.objects.filter(
+            user=request.user,
+            author=get_object_or_404(User, username=username)).exists():
+        Follow.objects.filter(
+            user=request.user,
+            author=get_object_or_404(User, username=username)).delete()
+        return redirect('posts:profile', username)
